+++ src/App.tsx
@@ -9,7 +9,7 @@ import FunitureSelection from "./pages/onboarding/FunitureSelection"
 import StyleSelection from "./pages/onboarding/StyleSelection"
 import FeaturedProduct from "./pages/onboarding/FeaturedProduct"
 import FeaturedBrand from "./pages/onboarding/FeaturedBrand"
-import Like from "./pages/like"
+import Like from "./pages/like/Like"
 import SignIn from "./pages/auth/SignIn"
 import SignUp from "./pages/auth/SignUp"
 import Cart from "./pages/cart"
