+++ src/pages/like/Like.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react"
-import { Box, Tabs, Tab, Typography } from "@mui/material"
+import { Box, Tabs, Tab } from "@mui/material"
 import NavigationBar from "../../components/NavigationBar"
 import ProductList from "./ProductList"
 import BrandList from "./BrandList"
