+++ src/pages/shop/ProductView/index.tsx
@@ -2,13 +2,13 @@ import React, { useState } from "react"
 import { useNavigate } from "react-router-dom"
 import { Tabs, Tab, Box, IconButton, Divider, Typography, SwipeableDrawer, MenuItem } from "@mui/material"
 import Select, { SelectChangeEvent } from "@mui/material/Select"
-import { NaviWrap } from "../../components/navigationbar.styles"
-import { SecondaryButton, PrimaryButton } from "../../styles/buttons.styles"
-import { ProductMainInfo, BrandInfo, OrderButton, ProductViewTabs, ViewTitle } from "./productView.styles"
-import { ArrowRight, IconHeartSmall, IconLikeOff, IconLikeOn } from "../../assets/images"
-import { sofa01 } from "../../assets/images/product"
-import { alfdn } from "../../assets/images/brand"
-import NoTitle from '../../components/title/NoTitle'
+import { NaviWrap } from "../../../components/navigationbar.styles"
+import { SecondaryButton, PrimaryButton } from "../../../styles/buttons.styles"
+import { ProductMainInfo, BrandInfo, OrderButton, ProductViewTabs, ViewTitle } from "./index.styles"
+import { ArrowRight, IconHeartSmall, IconLikeOff, IconLikeOn } from "../../../assets/images"
+import { sofa01 } from "../../../assets/images/product"
+import { alfdn } from "../../../assets/images/brand"
+import NoTitle from '../../../components/title/NoTitle'
 import DetailViewPanel from "./DetailViewPanel"
 import ReviewPanel from "./ReviewPanel"
 import PersonalQna from "./PersonalQna"
